+++ src/fiber_asm.c
@@ -17,7 +17,7 @@
 */
 
 #include "php_async.h"
-#include "fiber_stack.h"
+#include "async_stack.h"
 
 #undef ASM_CALLDECL
 #if (defined(i386) || defined(__i386__) || defined(__i386) \
