+++ src/task_scheduler.c
@@ -18,6 +18,8 @@
 
 #include "php_async.h"
 
+#include "async_task.h"
+
 ZEND_DECLARE_MODULE_GLOBALS(async)
 
 static zend_class_entry *async_task_scheduler_ce;
@@ -162,34 +164,32 @@ zend_bool async_task_scheduler_enqueue(async_task *task)
 		return 0;
 	}
 
-	if (!scheduler->dispatching) {
-		if (UNEXPECTED(EG(exception))) {
-			if (task->fiber.status == ASYNC_FIBER_STATUS_INIT) {
-				ZVAL_OBJ(&task->result, EG(exception));
-				EG(exception) = NULL;
+	if (!scheduler->dispatching && UNEXPECTED(EG(exception))) {
+		if (task->fiber.status == ASYNC_FIBER_STATUS_INIT) {
+			ZVAL_OBJ(&task->result, EG(exception));
+			EG(exception) = NULL;
 
-				zend_clear_exception();
+			zend_clear_exception();
 
-				zend_fcall_info_args_clear(&task->fiber.fci, 1);
-				zval_ptr_dtor(&task->fiber.fci.function_name);
+			zend_fcall_info_args_clear(&task->fiber.fci, 1);
+			zval_ptr_dtor(&task->fiber.fci.function_name);
 
-				task->operation = ASYNC_TASK_OPERATION_NONE;
-				task->fiber.status = ASYNC_FIBER_STATUS_FAILED;
+			task->operation = ASYNC_TASK_OPERATION_NONE;
+			task->fiber.status = ASYNC_FIBER_STATUS_FAILED;
 
-				async_awaitable_trigger_continuation(&task->continuation, &task->result, 0);
-			} else if (task->fiber.status == ASYNC_FIBER_STATUS_SUSPENDED) {
-				ZVAL_OBJ(&task->error, EG(exception));
-				EG(exception) = NULL;
+			async_awaitable_trigger_continuation(&task->continuation, &task->result, 0);
+		} else if (task->fiber.status == ASYNC_FIBER_STATUS_SUSPENDED) {
+			ZVAL_OBJ(&task->error, EG(exception));
+			EG(exception) = NULL;
 
-				zend_clear_exception();
+			zend_clear_exception();
 
-				async_task_dispose(task);
-			}
-
-			OBJ_RELEASE(&task->fiber.std);
-
-			return 0;
+			async_task_dispose(task);
 		}
+
+		OBJ_RELEASE(&task->fiber.std);
+
+		return 0;
 	}
 
 	if (!scheduler->dispatching && scheduler->ready.first == NULL) {
